requirements.txt
configargparse
joblib==1.2.0
matplotlib==3.6.2
scipy==1.9.3
tensorboard==2.11.0
torchopt
timm
tqdm
wandb
python-dotenv
pandas
ipython
git+https://github.com/openai/CLIP.git